+++ .github/workflows/ci.yaml
@@ -36,7 +36,7 @@ jobs:
 
       - name: Test with pytest
         run: |
-          python -m pytest --cov=smart_extract --cov-report=xml
+          python -m pytest --cov=smart_fetch --cov-report=xml
 
       - name: Log missing coverage
         run: |
