+++ .github/workflows/pypi.yaml
@@ -23,7 +23,7 @@ jobs:
     - name: Set version from tag
       run: |
         VERSION=$(echo $GITHUB_REF_NAME | sed 's/^v//')
-        sed -i "s/1\!0\.0\.0/$VERSION/w changes.txt" smart_extract/__init__.py
+        sed -i "s/1\!0\.0\.0/$VERSION/w changes.txt" smart_fetch/__init__.py
         [ -s changes.txt ] || exit 1  # validate that we successfully set the version
 
     - name: Build
