+++ README.md
@@ -1,4 +1,4 @@
-# SMART Extract
+# SMART Fetch
 
 A tool for extracting FHIR data from your EHR.
 
@@ -7,20 +7,20 @@ A tool for extracting FHIR data from your EHR.
 If you want to skip any description and dive right in:
 
 ```shell
-pipx install smart-extract
+pipx install smart-fetch
 
-smart-extract export \
+smart-fetch export \
   --fhir-url https://bulk-data.smarthealthit.org/fhir \
   /tmp/export
 ```
 
 After running the above, you'll have a bunch of NDJSON resource files sitting in `/tmp/export`.
 
-Run `smart-extract --help` to learn more about the various options.
+Run `smart-fetch --help` to learn more about the various options.
 
 ## Initial Export
 
-SMART Extract can do an initial pull from your EHR,
+SMART Fetch can do an initial pull from your EHR,
 with either a "bulk export" or by manually searching through resources 
 based on a list of patient MRNs,
 when bulk exporting is either not available or too slow.
@@ -31,7 +31,7 @@ Very often, the initial export from the EHR will be missing some useful data.
 There will be referenced Medications, Observations, and clinical notes that you need to manually
 download to have a complete set of data.
 
-SMART Extract can help fill in those gaps by post-processing (or hydrating) your FHIR data.
+SMART Fetch can help fill in those gaps by post-processing (or hydrating) your FHIR data.
 
 Mostly these hydration tasks focus on grabbing data now so that you'll always have it,
 no matter what happens with your EHR later.
@@ -49,22 +49,22 @@ no matter what happens with your EHR later.
 
 ## Managing Exports Over Time
 
-Extracting data for a group is rarely as simple as a single export.
+Fetching data for a group is rarely as simple as a single export.
 You may want to export some resources separately, for performance reasons
 (e.g. Observations can take a long time).
 Or you may want to grab updates to the group's data over time.
 
-SMART Extract can help with that, by keeping each export in its own subfolder
+SMART Fetch can help with that, by keeping each export in its own subfolder
 for individual processing,
 but also pooling all the results together with symlinks for convenience.
 
 ### Example
 
 ```shell
-pipx install smart-extract
+pipx install smart-fetch
 
 # Initial export
-smart-extract export \
+smart-fetch export \
   --fhir-url https://bulk-data.smarthealthit.org/fhir \
   --type Encounter,Patient \
   /tmp/export
@@ -75,7 +75,7 @@ ls -l /tmp/export
 # Patient.001.ndjson.gz -> 001.2025-06-26/Patient.001.ndjson.gz
 
 # Second export with a --since date, a new resource, and a nickname for the export
-smart-extract export \
+smart-fetch export \
   --fhir-url https://bulk-data.smarthealthit.org/fhir \
   --type Encounter,Patient,Condition \
   --since 2020-01-01 \
