--- smart_extract/__init__.py
@@ -1,3 +1,0 @@
-"""Extract FHIR data from your EHR easily and completely"""
-
-__version__ = "1!0.0.0"
