+++ smart_fetch/__init__.py
@@ -1,0 +1,3 @@
+"""Fetch FHIR data from your EHR easily and completely"""
+
+__version__ = "1!0.0.0"
