+++ smart_fetch/iter_utils.py
@@ -9,7 +9,7 @@ from typing import TypeVar
 import cumulus_fhir_support as cfs
 import rich.progress
 
-from smart_extract import cli_utils, ndjson, timing
+from smart_fetch import cli_utils, ndjson, timing
 
 Item = TypeVar("Item")
