+++ smart_fetch/ndjson.py
