+++ smart_fetch/resources.py
