+++ tests/test_hydrate_obs.py
@@ -2,7 +2,7 @@ import os
 
 import httpx
 
-from smart_extract import resources
+from smart_fetch import resources
 from tests import utils
